+++ packages/pangu/pangu/src/helper/index.ts
@@ -1,4 +1,8 @@
-import Title from "./title";
-import Loading from "./loading";
+import Title from './title';
+import Loading from './loading';
+import { ErrorLog } from './error';
 
-export  { Title, Loading };
+export * from './query';
+export * from './define';
+
+export { Title, Loading, ErrorLog };