+++ packages/pangu/pangu/src/hook/useQuery.ts
@@ -1,6 +1,6 @@
 import { useMemo } from 'react';
 import { useLocation } from 'react-router-dom';
-import { queryParse } from '../helper/query';
+import { queryParse } from '../helper';
 
 const useQuery = (): Record<string, unknown> => {
     const location = useLocation();