+++ packages/pangu/pangu/src/index.ts
@@ -1,8 +1,10 @@
 import PanGu from './core';
 
+export default PanGu;
+
 export * from 'react-router-dom';
 export * from './hoc';
 export * from './hook';
+export * from './helper';
 export type { PluginType, RoutePlugin } from './plugin';
 export type { PanGuRouteObject, RouterModeType, RouterProps } from './router';
-export default PanGu;